requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
Deprecated==1.2.13
Django==4.0.6
django-channels==0.7.0
django-redis==4.12.1
djangorestframework==3.13.1
idna==3.4
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dotenv==0.19.0
pytz==2022.1
redis>=4.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.14
websocket-client==1.4.2
wrapt==1.14.1
